modernize the build tool and use Minilla.
The biggest benefit for this is to support test requires in
META.json. Module::Install doesn't differentiate test_requires and
build_requires, so all the test-only modules are installed during the
installation.

For now, this change uses Minilla, but if you add one dist.ini file
it is easy to switch to Dist::Milla too :)